my_app/app_requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
exceptiongroup==1.1.1
Flask==2.3.2
flask-paginate==2022.1.8
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
flask-talisman==1.1.0
pyOpenSSL==24.1.0
